requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
decorator==4.4.2
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestql==1.0.1
imgkit==1.0.2
lxml==4.6.1
PyPDF2==1.26.0
python-decouple==3.3
pytz==2020.1
six==1.14.0
soupsieve==2.0.1
sqlparse==0.4.1
validators==0.18.1
